JPERF-730: Refactor "Formulas", so that Standalone matches DC and it's more explicit which IP addresses are used for what
This also prepares for further changes around JPERF-730
